요구사항 1 ~ 6 구현

+++ src/main/java/com/preonboarding/backend/user/model/vo/UserId.java
@@ -15,5 +15,5 @@ import java.io.Serializable;
 public class UserId implements Serializable {
 
     @Column(name = "user_id")
-    private int seq;
+    private long seq;
 }
